--- Utilities/logfile.txt
@@ -1,28 +1,0 @@
-2024-11-15 03:48:55: Starting email search...
-2024-11-15 03:50:17: Starting email search...
-2024-11-15 03:50:17: Email Searching Done
-2024-11-15 03:50:17: The result has been saved to email.json
-2024-11-15 05:06:59: Starting email search...
-2024-11-15 05:07:50: Starting email search...
-2024-11-15 05:07:50: Email Searching Done
-2024-11-15 05:07:50: The result has been saved to email.json
-2024-11-15 05:08:25: Starting email search...
-2024-11-15 05:08:25: Email Searching Done
-2024-11-15 05:08:25: The result has been saved to email.json
-2024-11-15 05:08:29: Starting email search...
-2024-11-15 05:08:29: Email Searching Done
-2024-11-15 05:08:29: The result has been saved to email.json
-2024-11-15 05:09:51: Starting email search...
-2024-11-15 05:09:51: Email Searching Done
-2024-11-15 05:09:51: The result has been saved to email.json
-2024-11-15 05:10:24: Starting email search...
-2024-11-15 05:11:53: Starting email search...
-2024-11-15 05:13:41: Starting email search...
-2024-11-15 05:14:46: Starting email search...
-2024-11-15 05:15:46: Starting email search...
-2024-11-15 05:15:51: Email Searching Done
-2024-11-15 05:15:51: The result has been saved to email.json
-2024-11-15 05:26:43: Starting email search...
-2024-11-15 05:26:48: Email Searching Done
-2024-11-15 05:26:48: The result has been saved to email.json
-2024-11-16 01:58:17: Starting GitHub search...
